+++ packages/schema-blocks/src/core/Definition.ts
@@ -1,9 +1,12 @@
 import { BlockValidation, BlockValidationResult } from "./validation";
 import { BlockInstance } from "@wordpress/blocks";
+import { BlockPresence } from "./validation/BlockValidationResult";
 import { isArray, mergeWith } from "lodash";
 import Instruction from "./Instruction";
 import Leaf from "./Leaf";
 import logger from "../functions/logger";
+import validateMany from "../functions/validators/validateMany";
+import { getAllDescendantIssues } from "../functions/validators";
 
 export type DefinitionClass<T extends Definition> = {
 	new( separator: string, template?: string, instructions?: Record<string, Instruction>, tree?: Leaf ): T;
@@ -80,17 +83,28 @@ export default abstract class Definition {
 			return null;
 		}
 
-		const validation = new BlockValidationResult( blockInstance.clientId, blockInstance.name, BlockValidation.Unknown );
+		const validation = new BlockValidationResult( blockInstance.clientId, blockInstance.name, BlockValidation.Unknown, BlockPresence.Unknown );
+
+		logger.startGroup( `Validation results: ${ blockInstance.name }` );
 
 		validation.issues = Object.values( this.instructions ).map( instruction => {
 			const issue = instruction.validate( blockInstance );
-			logger.debug( "validating " + instruction.options.name + " against " + blockInstance.name + " => " + BlockValidation[ issue.result ] );
+			logger.debug( `${ instruction.options.name} => ${ BlockValidation[ issue.result ] }` );
 			return issue;
 		} ).filter( issue => issue.result !== BlockValidation.Skipped );
 
-		// In case of a block with just one innerblock, this prevents a duplicate, identical wrapper.
+		logger.endGroup();
+
+		// In case of a block with just one validated (i.e., not skipped) instruction, this prevents a duplicate, identical wrapper.
 		if ( validation.issues.length === 1 ) {
 			return validation.issues[ 0 ];
+			// eslint-disable-next-line no-inline-comments,brace-style
+		} // If a block has multiple validated instructions, we should `validateMany` to avoid returning a BlockValidation.Unknown (-1).
+		else if ( validation.issues.length > 1 ) {
+			const descendantIssues = getAllDescendantIssues( validation );
+			// The first of the descendant issues is the parent, so we exclude it from the slice to avoid an infinite loop.
+			validation.issues = descendantIssues.slice( 1 );
+			return validateMany( validation );
 		}
 
 		return validation;