+++ packages/schema-blocks/src/functions/validators/index.ts
@@ -1,11 +1,11 @@
 import attributeExists from "./attributeExists";
 import attributeNotEmpty from "./attributeNotEmpty";
+import getAllDescendantIssues from "./getAllDescendantIssues";
 import getInvalidInnerBlocks from "./innerBlocksValid";
-import isValidResult from "./isValidResult";
 
 export {
 	attributeExists,
 	attributeNotEmpty,
+	getAllDescendantIssues,
 	getInvalidInnerBlocks,
-	isValidResult,
 };